refactor: move GenerateSourceRequest classes to PropertyBitPack.SourceGen.Models
Relocated GenerateSourceRequest and related classes to PropertyBitPack.SourceGen.Models.GenerateSourceRequest namespace. Updated all references and removed folder inclusion from PropertyBitPack.SourceGen.csproj.
