+++ PropertyBitPack.SourceGen/BitFieldPropertyAggregators/ReadOnlyAggregator.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
 using PropertyBitPack.SourceGen.Collections;
 using PropertyBitPack.SourceGen.Models;
+using PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
 using System;
 using System.Collections.Generic;
 using System.Text;
