+++ PropertyBitPack.SourceGen/Models/GenerateSourceRequest/UnnamedFieldGsr.cs
@@ -4,7 +4,7 @@ using System.Collections.Immutable;
 using System.Runtime.CompilerServices;
 using System.Text;
 
-namespace PropertyBitPack.SourceGen.Models;
+namespace PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
 internal sealed class UnnamedFieldGsr(NonExistingFieldRequest fieldRequest, ImmutableArray<BitFieldPropertyInfoRequest> properties) : NonExistingFieldGsr
 {
     public NonExistingFieldRequest FieldRequest => Unsafe.As<NonExistingFieldRequest>(Fields[0]);
