+++ PropertyBitPack.SourceGen/PropertiesSyntaxGenerators/ExistingFieldPropertiesGenerator.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Text;
+using PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
 
 namespace PropertyBitPack.SourceGen.PropertiesSyntaxGenerators;
