+++ PropertyBitPack.SourceGen/PropertiesSyntaxGenerators/IPropertiesSyntaxGenerator.cs
@@ -2,6 +2,7 @@
 using Microsoft.CodeAnalysis.Text;
 using PropertyBitPack.SourceGen.Collections;
 using PropertyBitPack.SourceGen.Models;
+using PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
