+++ PropertyBitPack.SourceGen/PropertiesSyntaxGenerators/PropertySyntaxGenerator.cs
@@ -1,4 +1,5 @@
 ﻿using PropertyBitPack.SourceGen.Models;
+using PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
 using System;
 using System.Collections.Generic;
 using System.Text;
