+++ PropertyBitPack.SourceGen/PropertyBitPackSourceGenerator.cs
@@ -13,6 +13,7 @@ using PropertyBitPack.SourceGen.AttributeParsers;
 using PropertyBitPack.SourceGen.BitFieldPropertyParsers;
 using PropertyBitPack.SourceGen.BitFieldPropertyAggregators;
 using PropertyBitPack.SourceGen.PropertiesSyntaxGenerators;
+using PropertyBitPack.SourceGen.Models.GenerateSourceRequest;
 
 namespace PropertyBitPack.SourceGen;
